refactor(acquisition): fix 'acqusition' typo
