+++ smac/acquisition/function/abstract_acquisition_function.py
@@ -50,7 +50,7 @@ class AbstractAcquisitionFunction:
 
         This method will be called after fitting the model, but before maximizing the acquisition
         function. As an examples, EI uses it to update the current fmin. The default implementation only updates the
-        attributes of the acqusition function which are already present.
+        attributes of the acquisition function which are already present.
 
         Calls `_update` to update the acquisition function attributes.
